+++ Sources/MapboxCoreNavigation/RerouteControllerDelegate.swift
@@ -4,8 +4,11 @@ import MapboxNavigationNative
 
 
 protocol ReroutingControllerDelegate: AnyObject {
-    // TODO: fill with Native RerouteController and Alternative routes integration
-    func rerouteControllerDidDetectReroute(_ rerouteController: RerouteController)
+    func rerouteControllerWantsSwitchToAlternative(_ rerouteController: RerouteController,
+                                                   response: RouteResponse,
+                                                   routeIndex: Int,
+                                                   options: RouteOptions)
+    func rerouteControllerDidDetectReroute(_ rerouteController: RerouteController) -> Bool
     func rerouteControllerDidRecieveReroute(_ rerouteController: RerouteController, response: RouteResponse, options: RouteOptions)
     func rerouteControllerDidCancelReroute(_ rerouteController: RerouteController)
     func rerouteControllerDidFailToReroute(_ rerouteController: RerouteController, with error: DirectionsError)