+++ Tests/MapboxCoreNavigationTests/MapboxCoreNavigationTests.swift
@@ -33,6 +33,7 @@ class MapboxCoreNavigationTests: TestCase {
     }
     
     override func tearDown() {
+        Navigator.shared.rerouteController.reroutesProactively = true
         super.tearDown()
         navigation = nil
         UserDefaults.resetStandardUserDefaults()
@@ -90,6 +91,7 @@ class MapboxCoreNavigationTests: TestCase {
                                              customRoutingProvider: MapboxRoutingProvider(.offline),
                                              credentials: Fixture.credentials,
                                              simulating: .never)
+        Navigator.shared.rerouteController.reroutesProactively = false
         
         // Coordinates from first step
         let coordinates = route.legs[0].steps[0].shape!.coordinates
@@ -141,6 +143,7 @@ class MapboxCoreNavigationTests: TestCase {
                                                         credentials: Fixture.credentials,
                                                         locationSource: locationManager,
                                                         simulating: .never)
+        Navigator.shared.rerouteController.reroutesProactively = false
         
         var receivedSpokenInstructions: [String] = []
         
@@ -198,6 +201,7 @@ class MapboxCoreNavigationTests: TestCase {
                                              credentials: Fixture.credentials,
                                              locationSource: locationManager,
                                              simulating: .never)
+        Navigator.shared.rerouteController.reroutesProactively = false
         
         expectation(forNotification: .routeControllerDidPassSpokenInstructionPoint,
                     object: navigation.router) { (notification) -> Bool in
@@ -340,6 +344,7 @@ class MapboxCoreNavigationTests: TestCase {
                                              customRoutingProvider: MapboxRoutingProvider(.offline),
                                              credentials: Fixture.credentials,
                                              locationSource: locationManager)
+        Navigator.shared.rerouteController.reroutesProactively = false
         navigation.router.refreshesRoute = false
         
         struct InstructionPoint {
@@ -474,6 +479,69 @@ class MapboxCoreNavigationTests: TestCase {
         navigation.router.reroute(from: CLLocation(latitude: 0, longitude: 0), along: navigation.router.routeProgress)
         
         waitForExpectations(timeout: 2) { (error) in
+            XCTAssertNil(error)
+        }
+    }
+    
+    func testNoUpdatesAfterFinishingRouting() {
+        navigation = MapboxNavigationService(routeResponse: response,
+                                             routeIndex: 0,
+                                             routeOptions: routeOptions,
+                                             customRoutingProvider: MapboxRoutingProvider(.offline),
+                                             credentials: Fixture.credentials,
+                                             simulating: .never)
+        
+        // Coordinates from first step
+        let coordinates = route.legs[0].steps[0].shape!.coordinates
+        let now = Date()
+        let locations = coordinates.enumerated().map {
+            CLLocation(coordinate: $0.element,
+                       altitude: -1,
+                       horizontalAccuracy: 10,
+                       verticalAccuracy: -1,
+                       course: -1,
+                       speed: 10,
+                       timestamp: now + $0.offset)
+        }
+        
+        var hasFinishedRouting = false
+        expectation(forNotification: .routeControllerProgressDidChange, object: navigation.router) { _ in
+            return hasFinishedRouting
+        }.isInverted = true
+        expectation(forNotification: .routeControllerWillReroute, object: navigation.router) { _ in
+            return hasFinishedRouting
+        }.isInverted = true
+        expectation(forNotification: .routeControllerDidReroute, object: navigation.router) { _ in
+            return hasFinishedRouting
+        }.isInverted = true
+        
+        let routerDelegateSpy = RouterDelegateSpy()
+        navigation.router.delegate = routerDelegateSpy
+        
+        routerDelegateSpy.onShouldDiscard = { _ in
+            if hasFinishedRouting {
+                XCTFail("Location updates should not be tracked.")
+            }
+            return false
+        }
+        
+        navigation.start()
+        
+        for location in locations {
+            navigation.locationManager(navigation.locationManager, didUpdateLocations: [location])
+            
+            if !hasFinishedRouting {
+                navigation.router.finishRouting()
+                hasFinishedRouting = true
+                
+                navigation.updateRoute(with: IndexedRouteResponse(routeResponse: response,
+                                                                  routeIndex: 0),
+                                       routeOptions: nil,
+                                       completion: nil)
+            }
+        }
+        
+        waitForExpectations(timeout: waitForInterval) { (error) in
             XCTAssertNil(error)
         }
     }